analytics package and button fixes
