+++ apps/framework-docs/src/pages/ingest-data/_meta.json
@@ -1,4 +1,4 @@
 {
-  "ingest-api": "via Ingestion API Enpoints",
+  "ingest-api": "via Ingestion API Endpoints",
   "ingest-sdk": "via Ingestion SDKs"
 }
