+++ apps/framework-landing/app/page.tsx
@@ -1,79 +1,19 @@
-import { ReactNode } from "react";
-
-import { VariantProps } from "class-variance-authority";
+import { FromToSection } from "./sections/home/FromToSection";
 import { FooterSection } from "./sections/FooterSection";
 import { HeroSection } from "./sections/home/HeroSection";
 import { PrimitivesCode } from "./sections/home/PrimitivesCode";
 import { FeaturesSection } from "./sections/home/FeaturesSection";
 import { DemoSection } from "./sections/home/DemoSection";
-
 import { SecondaryCTASection } from "./sections/home/HostWithBorealSection";
 import { cn } from "@514labs/design-system-components/utils";
-
-import {
-  Button,
-  buttonVariants,
-} from "@514labs/design-system-components/components";
-
-import { Text } from "@514labs/design-system-components/typography";
-import React from "react";
 import { WhatIsMooseFor } from "./sections/home/WhatIsMooseForV2";
-
-export const CTAText = ({
-  className,
-  children,
-}: {
-  className?: string;
-  children: ReactNode;
-}) => {
-  return (
-    <div
-      className={cn(
-        "text-center md:text-start text-primary text-4xl bg-muted rounded-md py-5 px-10 text-nowrap",
-        className,
-      )}
-    >
-      {children}
-    </div>
-  );
-};
-
-export interface CTAButtonProps extends VariantProps<typeof buttonVariants> {
-  className?: string;
-  children: ReactNode;
-  onClick?: () => void;
-}
-
-export const CTAButton = ({
-  className,
-  children,
-  variant,
-  onClick,
-}: CTAButtonProps) => {
-  return (
-    <Button
-      size={"default"}
-      variant={variant}
-      className={cn("border-primary w-full sm:w-auto", className)}
-      onClick={onClick}
-    >
-      <Text
-        className={cn(
-          variant === "outline" ? "text-primary" : "text-primary-foreground",
-        )}
-      >
-        {children}
-      </Text>
-    </Button>
-  );
-};
 
 export const CTABar = ({
   className,
   children,
 }: {
   className?: string;
-  children: ReactNode;
+  children: React.ReactNode;
 }) => {
   return (
     <div className={cn("flex flex-col md:flex-row gap-5", className)}>
@@ -85,7 +25,8 @@ export const CTABar = ({
 export default function Home() {
   return (
     <main>
-      <HeroSection />
+      <HeroSection className="max-w-5xl mx-auto flex flex-col items-center px-5 my-16 sm:my-32 py-10 pb-20" />
+      <FromToSection />
       <DemoSection />
       <PrimitivesCode />
       <WhatIsMooseFor />
