+++ packages/design-system-components/components/CTAs.tsx
@@ -1,6 +1,7 @@
 import { ReactNode } from "react";
 import { VariantProps } from "class-variance-authority";
 import { cn } from "../lib/utils";
+
 import { Text } from "./typography/standard";
 import { Button, buttonVariants } from "./ui/button";
